When adding every track in simple-mpc-query-mode don't forget last one.
This happens because when adding things normally the line point is on
isn't actually added. This is by design because often times you just do
the following when selecting a number of tracks:

<C-SPC>
<C-n>, <C-n>, ...
<RET>

After this, point is on column 0 on a certain track. Because of how
Emacs' highlighting works however, it doesn't really look like you
selected it. So to take that into account we subtract one from the line
`region-end' is on. This way the tracks you add correspond nicely with
the highlighted lines. I'm pretty sure this is also how it works in
Mingus, but I'm not sure.

Anyway, this just checks if `region-end' is on `point-max' to check to
include the current-line.