Respect simple-mpc-playlist-format in simple-mpc-query
Before this there was no format given to mpc when searching. By default
mpc then uses the %file% metadata delimiter. This is fine if you use MPD
and you have organized, descriptive filenames. Mopidy however changes
the format of this delimiter to be:

urlencode(backend:type:identifier)

Because for e.g. Spotify results there is no real 'file'. Instead there
is a unique identifier. This unfortunately meant that although searching
worked with Mopidy as a server instead of MPD, you ended up with a query
buffer looking like:

...
spotify:artist:0fDF0jjmdouCIeWhNnblwV
spotify:artist:5EWpmduD1uxFH0D9uYVyJg
spotify:artist:6khZGp3Cq9nGdWnYUwJGxu
spotify:album:0LBQdWnuV0CAXyPIngb0UX
spotify:album:0TNvR6Co87E75jWohyn2jq
spotify:album:52SHWfP9ggb7cg1E0iBH1h
...

which isn't very useful.

To solve this problem we pass the user-defined
simple-mpc-playlist-format to 'mpc search'. The main issue with
this (and the main reason this was not done earlier) is that then you
cannot directly pass the selected lines in the query buffer to 'mpc
add'. For the add command to work correctly you have to pass the %file%
of the result.

To allow for this simple-mpc-query-current-result-alist was
introduced. It contains a (simple-mpc-playlist-format . %file%) pair for
every result in the query. This way we can look up %file% for whatever
the user selects.

To accomplish this in one 'mpc search' we prepend the %file% metadata
delimiter to simple-mpc-playlist-format, delimited by (simple-mpc). We
then extract out of the results the (simple-mpc-playlist-format
. %file%) pairs for simple-mpc-query-current-result-alist.

If you happen to have two results with the exact same name in your query
buffer this approach won't work, but hopefully that shouldn't happen.