Output playlists as a table
It is nice to have the playlist formatted as a table, otherwise the
results can be a bit hard to read. The previously introduced
simple-mpc-query-current-result-alist allows us to do this relatively
easily.

This introduces simple-mpc-table-separator, a variable that can be set
by the user (nil by default). When this variable is set the column
program will be used to format the output from mpc as a table.

This requires a slight change to how we did the file metadata
delimiter. We now append it instead of prepending it. An example of why
this is necessary:

(s)spotify:artist:abc(s)                     title
(s)spotify:track:a1b2c3(s)artist    album    title

Whatever is between the separators has a variable length. When you
remove it you'll end up with unaligned text. By appending instead of
prepending we avoid this issue.