Handle empty simple-mpc-playlist-format in simple-mpc-query
89440b017d74fd675f22e0505ba8d86eabf0bc5a introduced a bug which occurs
when you do not set simple-mpc-playlist-format. It will then have it's
default value of "".

In this case you will end up with the query view displaying all your
results as empty strings. Everything will still work, so you can still
add/remove files to the current playlist, but the entire buffer will
appear blank. This issue does not occur in simple-mpc-current-playlist
because there we only pass the format parameter if
simple-mpc-playlist-format is set. In simple-mpc-query however, we
always pass format because of what was done in
89440b017d74fd675f22e0505ba8d86eabf0bc5a.

We use this opportunity to make the behavior of passing the format
parameter more consistent between simple-mpc-query and
simple-mpc-current-playlist. Now we always pass the format parameter in
both modes, opting to pass "%file%" if simple-mpc-playlist-format is not
set.

Fixes #4