Don't depend on non-portable -n column flag
Old versions of column merge multiple adjacent delimiters. This can be
disabled by the -n flag in old Debian GNU/Linux versions of the column
program.

However in new versions of column (at least the one included in
util-linux >=2.31.1) the -n flag is used to specify a table name for
JSON output. This breaks simple-mpc's table formatting.

Instead we get rid of the -n flag (it was not portable anyway) and
make sure there are no empty columns by inserting slashes (/) into
them.