Implement volume control
This allows controlling the volume of mpd or mopidy through
simple-mpc.

The volume step size is defined in `simple-mpc-volume-step-size' and
is a customizable variable.

Since the volume parameter is handled the same way as the seek
parameter `simple-mpc-convert-number-to-relative-string' was extracted
so it can be reused for both `simple-mpc-seek-internal' and
`simple-mpc-modify-volume-internal'.